5/20/24
I realized that I completely forgot about the custom cursor plan.. Maybe tomorrow I’ll work on it. Now that my objective for this project has changed, I will need to remove the background I added and figure out a way to implement my idea.

There was a bug with the moving text, so I figured out how to fix it and also I began to work on getting the button triggered, didn't finish.