5/31/24
I was able to fix the cursor finally, but there’s still a slight problem with it, one that might resurface later on, but I’ll figure it out when the time comes. 

I found a new background for the game, and over the weekend, I will create characters and add dialogue.